+++ keyring/src/error.rs
@@ -1,25 +1,22 @@
-use std::result;
 use thiserror::Error;
 use tokio::task::JoinError;
 
 /// The global `Result` alias of the library.
-pub type Result<T> = result::Result<T, Error>;
+pub type Result<T> = std::result::Result<T, Error>;
 
 /// The global `Error` enum of the library.
 #[derive(Debug, Error)]
 pub enum Error {
     #[error("cannot build keyring entry using key `{1}`")]
-    BuildEntryError(#[source] keyring_native::Error, String),
+    BuildEntryError(#[source] native::Error, String),
     #[error("cannot get secret from keyring matching `{1}`")]
-    GetSecretError(#[source] keyring_native::Error, String),
+    GetSecretError(#[source] native::Error, String),
     #[error("cannot find secret from keyring matching `{1}`")]
-    FindSecretError(#[source] keyring_native::Error, String),
+    FindSecretError(#[source] native::Error, String),
     #[error("cannot set secret from keyring matching `{1}`")]
-    SetSecretError(#[source] keyring_native::Error, String),
+    SetSecretError(#[source] native::Error, String),
     #[error("cannot delete secret from keyring matching `{1}`")]
-    DeleteSecretError(#[source] keyring_native::Error, String),
-    #[error("cannot build keyutils credentials using key {1}")]
-    BuildCredentialsError(#[source] keyring_native::Error, String),
+    DeleteSecretError(#[source] native::Error, String),
 
     #[error(transparent)]
     JoinError(#[from] JoinError),